requirements.txt
aiohttp==3.9.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fastapi==0.104.1
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.1
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
lapx==0.9.2
loguru==0.7.2
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.7.0
networkx==3.6
numpy==1.24.3
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==25.0
pandas==2.3.3
Pillow==10.1.0
propcache==0.4.1
psutil==7.1.3
py-cpuinfo==9.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.1
requests==2.32.5
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.14.0
thop==0.1.1.post2209072238
torch==2.1.0
torchvision==0.16.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.0.220
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.1
websockets==15.0.1
win32_setctime==1.2.0
yarl==1.22.0
